@@ -5,8 +5,10 @@ import com.vgaw.scaffolddemo.data.user.UserInfo;
 import java.util.ArrayList;
 import java.util.List;
 
+import io.reactivex.Observable;
+
 public class MockUtil {
-    public static List<UserInfo> buildUserList() {
+    public static Observable<List<UserInfo>> buildUserList() {
         List<UserInfo> list = new ArrayList<>();
 
         UserInfo userInfo = new UserInfo();
@@ -39,22 +41,14 @@ public class MockUtil {
         userInfo4.setId("meiwen");
         list.add(userInfo4);
 
-        return list;
+        return Observable.just(list);
     }
 
-    public static UserInfo buildLoginSucData() {
-        UserInfo userInfo = new UserInfo();
-        userInfo.setAvatar("https://ss0.bdstatic.com/70cFuHSh_Q1YnxGkpoWK1HF6hhy/it/u=2196657174,1994863858&fm=26&gp=0.jpg");
-        userInfo.setName("洛华");
-        userInfo.setId("luohua");
-        return userInfo;
-    }
-
-    public static List<String> buildADList() {
+    public static Observable<List<String>> buildADList() {
         List<String> list = new ArrayList<>();
         list.add("https://ss0.bdstatic.com/70cFvHSh_Q1YnxGkpoWK1HF6hhy/it/u=672721443,3070988109&fm=26&gp=0.jpg");
         list.add("https://timgsa.baidu.com/timg?image&quality=80&size=b9999_10000&sec=1571744860835&di=a6df90fdb4aa96fa11aa9b47393a3048&imgtype=0&src=http%3A%2F%2Fhbimg.b0.upaiyun.com%2F7afa485b8ae44b419df9425106631c168a12bdf0172b1-b7EtqR_fw658");
         list.add("https://timgsa.baidu.com/timg?image&quality=80&size=b9999_10000&sec=1571744905068&di=8ce5d0f994ed74bea4360574c6c7f35d&imgtype=0&src=http%3A%2F%2Fhbimg.b0.upaiyun.com%2F3fef80e17c9e1e7c5fa8965af6a6800610eb50d22ee0-u9jiq0_fw658");
-        return list;
+        return Observable.just(list);
     }
 }
