@@ -6,11 +6,11 @@ import com.vgaw.scaffold.http.service.HttpService;
 import com.vgaw.scaffold.json.JsonUtil;
 
 import java.io.File;
-import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
 
+import io.reactivex.Observable;
 import okhttp3.MediaType;
 import okhttp3.MultipartBody;
 import okhttp3.OkHttpClient;
@@ -18,59 +18,55 @@ import okhttp3.RequestBody;
 import okhttp3.ResponseBody;
 import okhttp3.internal.proxy.NullProxySelector;
 import okhttp3.logging.HttpLoggingInterceptor;
-import retrofit2.Callback;
-import retrofit2.Response;
 import retrofit2.Retrofit;
+import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory;
 
 /**
  * Created by caojin on 2017/5/25.
+ * 不同地址的请求需使用不同的实例
  */
 
 public class HttpApi {
     private HttpConfig mHttpConfig;
 
+    private HttpService mService;
+    private HttpService mDownloadService;
+
     public void config(HttpConfig config) {
         mHttpConfig = config;
     }
 
-    /**
-     * small file download also use this
-     *
-     * @param path
-     * @param queryMap
-     * @return
-     */
-    public void get(OkHttpClient client, String path, Map<String, String> headerMap, Map<String, Object> queryMap, Callback<ResponseBody> callback) {
-        getService(client).get(path, headerMap, queryMap).enqueue(callback);
+    public Observable<ResponseBody> get(OkHttpClient client, String path, Map<String, String> headerMap, Map<String, Object> queryMap) {
+        return getService(client).get(path, headerMap, queryMap);
     }
 
-    public void get(OkHttpClient client, String path, Map<String, Object> queryMap, Callback<ResponseBody> callback) {
-        getService(client).get(path, queryMap).enqueue(callback);
+    public Observable<ResponseBody> get(OkHttpClient client, String path, Map<String, Object> queryMap) {
+        return getService(client).get(path, queryMap);
     }
 
-    public void post(OkHttpClient client, String path, Map<String, String> headerMap, Map<String, Object> fieldMap, Callback<ResponseBody> callback) {
-        getService(client).post(path, headerMap, fieldMap).enqueue(callback);
+    public Observable<ResponseBody> post(OkHttpClient client, String path, Map<String, String> headerMap, Map<String, Object> fieldMap) {
+        return getService(client).post(path, headerMap, fieldMap);
     }
 
-    public void post(OkHttpClient client, String path, Map<String, Object> fieldMap, Callback<ResponseBody> callback) {
-        getService(client).post(path, fieldMap).enqueue(callback);
+    public Observable<ResponseBody> post(OkHttpClient client, String path, Map<String, Object> fieldMap) {
+        return getService(client).post(path, fieldMap);
     }
 
-    public void postJson(OkHttpClient client, String path, Map<String, String> headerMap, Map<String, Object> fieldMap, Callback<ResponseBody> callback) {
+    public Observable<ResponseBody> postJson(OkHttpClient client, String path, Map<String, String> headerMap, Map<String, Object> fieldMap) {
         RequestBody requestBody =
                 RequestBody.create(MediaType.parse("application/json; charset=utf-8"),
                         JsonUtil.toJson(fieldMap));
-        getService(client).postJson(path, headerMap, requestBody).enqueue(callback);
+        return getService(client).postJson(path, headerMap, requestBody);
     }
 
-    public void postJson(OkHttpClient client, String path, Map<String, Object> fieldMap, Callback<ResponseBody> callback) {
+    public Observable<ResponseBody> postJson(OkHttpClient client, String path, Map<String, Object> fieldMap) {
         RequestBody requestBody =
                 RequestBody.create(MediaType.parse("application/json; charset=utf-8"),
                         JsonUtil.toJson(fieldMap));
-        getService(client).postJson(path, requestBody).enqueue(callback);
+        return getService(client).postJson(path, requestBody);
     }
 
-    public void upload(OkHttpClient client, String path, File file, Map<String, String> headerMap, Map<String, String> fieldMap, Callback<ResponseBody> callback) {
+    public Observable<ResponseBody> upload(OkHttpClient client, String path, File file, Map<String, String> headerMap, Map<String, String> fieldMap) {
         RequestBody requestFile = RequestBody.create(MediaType.parse("multipart/form-data"), file);
         MultipartBody.Part partFile = MultipartBody.Part.createFormData("file", file.getName(), requestFile);
         Map<String, RequestBody> partMap = new HashMap<>();
@@ -79,10 +75,10 @@ public class HttpApi {
             String value = fieldMap.get(key);
             partMap.put(key, RequestBody.create(MediaType.parse("multipart/form-data"), value));
         }
-        getService(client).upload(path, partFile, headerMap, partMap).enqueue(callback);
+        return getService(client).upload(path, partFile, headerMap, partMap);
     }
 
-    public void upload(OkHttpClient client, String path, File file, Map<String, String> headerMap, Map<String, String> fieldMap, OnProgressUpdateListener listener, Callback<ResponseBody> callback) {
+    public Observable<ResponseBody> upload(OkHttpClient client, String path, File file, Map<String, String> headerMap, Map<String, String> fieldMap, OnProgressUpdateListener listener) {
         RequestBody requestFile = new ProgressRequestBody(file, MediaType.parse("multipart/form-data"), listener);
         MultipartBody.Part partFile = MultipartBody.Part.createFormData("file", file.getName(), requestFile);
         Map<String, RequestBody> partMap = new HashMap<>();
@@ -91,94 +87,90 @@ public class HttpApi {
             String value = fieldMap.get(key);
             partMap.put(key, RequestBody.create(MediaType.parse("multipart/form-data"), value));
         }
-        getService(client).upload(path, partFile, headerMap, partMap).enqueue(callback);
+        return getService(client).upload(path, partFile, headerMap, partMap);
     }
 
     /**
-     * for big file downloading
+     * for big file downloading, small file downloading may use get method
      *
      * @return
      */
-    public void download(OkHttpClient client, String path, Map<String, Object> paramMap, OnProgressUpdateListener listener, Callback<ResponseBody> callback) {
-        getDownloadService(client, null, listener).download(path, paramMap).enqueue(callback);
+    public Observable<ResponseBody> download(OkHttpClient client, String path, Map<String, Object> paramMap, OnProgressUpdateListener listener) {
+        return getDownloadService(client, null, listener).download(path, paramMap);
     }
 
-    private void download(OkHttpClient client, String baseUrl, String path, Map<String, Object> paramMap, OnProgressUpdateListener listener, Callback<ResponseBody> callback) {
-        getDownloadService(client, baseUrl, listener).download(path, paramMap).enqueue(callback);
+    private Observable<ResponseBody> download(OkHttpClient client, String baseUrl, String path, Map<String, Object> paramMap, OnProgressUpdateListener listener) {
+        return getDownloadService(client, baseUrl, listener).download(path, paramMap);
     }
 
     /*without client version*/
 
-    public void get(String path, Map<String, String> headerMap, Map<String, Object> queryMap, Callback<ResponseBody> callback) {
-        get(null, path, headerMap, queryMap, callback);
+    public Observable<ResponseBody> get(String path, Map<String, String> headerMap, Map<String, Object> queryMap) {
+        return get(null, path, headerMap, queryMap);
     }
 
-    public void get(String path, Map<String, Object> queryMap, Callback<ResponseBody> callback) {
-        get(null, path, queryMap, callback);
+    public Observable<ResponseBody> get(String path, Map<String, Object> queryMap) {
+        return get(null, path, queryMap);
     }
 
-    public void post(String path, Map<String, String> headerMap, Map<String, Object> fieldMap, Callback<ResponseBody> callback) {
-        post(null, path, headerMap, fieldMap, callback);
+    public Observable<ResponseBody> post(String path, Map<String, String> headerMap, Map<String, Object> fieldMap) {
+        return post(null, path, headerMap, fieldMap);
     }
 
-    public void post(String path, Map<String, Object> fieldMap, Callback<ResponseBody> callback) {
-        post(null, path, fieldMap, callback);
+    public Observable<ResponseBody> post(String path, Map<String, Object> fieldMap) {
+        return post(null, path, fieldMap);
     }
 
-    public void postJson(String path, Map<String, String> headerMap, Map<String, Object> fieldMap, Callback<ResponseBody> callback) {
-        postJson(null, path, headerMap, fieldMap, callback);
+    public Observable<ResponseBody> postJson(String path, Map<String, String> headerMap, Map<String, Object> fieldMap) {
+        return postJson(null, path, headerMap, fieldMap);
     }
 
-    public void postJson(String path, Map<String, Object> fieldMap, Callback<ResponseBody> callback) {
-        postJson(null, path, fieldMap, callback);
+    public Observable<ResponseBody> postJson(String path, Map<String, Object> fieldMap) {
+        return postJson(null, path, fieldMap);
     }
 
-    public void upload(String path, File file, Map<String, String> headerMap, Map<String, String> fieldMap, Callback<ResponseBody> callback) {
-        upload(null, path, file, headerMap, fieldMap, callback);
+    public Observable<ResponseBody> upload(String path, File file, Map<String, String> headerMap, Map<String, String> fieldMap) {
+        return upload(null, path, file, headerMap, fieldMap);
     }
 
-    public void upload(String path, File file, Map<String, String> headerMap, Map<String, String> fieldMap, OnProgressUpdateListener listener, Callback<ResponseBody> callback) {
-        upload(null, path, file, headerMap, fieldMap, listener, callback);
+    public Observable<ResponseBody> upload(String path, File file, Map<String, String> headerMap, Map<String, String> fieldMap, OnProgressUpdateListener listener) {
+        return upload(null, path, file, headerMap, fieldMap, listener);
     }
 
-    public void download(String baseUrl, String path, Map<String, Object> paramMap, OnProgressUpdateListener listener, Callback<ResponseBody> callback) {
-        download(null, baseUrl, path, paramMap, listener, callback);
-    }
-
-    /*without callback version*/
-
-    public Response<ResponseBody> post(OkHttpClient client, String path, Map<String, String> headerMap, Map<String, Object> fieldMap) throws IOException {
-        return getService(client).post(path, headerMap, fieldMap).execute();
-    }
-
-    public Response<ResponseBody> post(OkHttpClient client, String path, Map<String, Object> fieldMap) throws IOException {
-        return getService(client).post(path, fieldMap).execute();
-    }
-
-    public Response<ResponseBody> get(OkHttpClient client, String path, Map<String, Object> queryMap) throws IOException {
-        return getService(client).get(path, queryMap).execute();
+    public Observable<ResponseBody> download(String baseUrl, String path, Map<String, Object> paramMap, OnProgressUpdateListener listener) {
+        return download(null, baseUrl, path, paramMap, listener);
     }
 
     private HttpService getService(OkHttpClient client) {
-        Retrofit.Builder builder = new Retrofit.Builder()
-                .baseUrl(getURL());
-        if (client == null && mHttpConfig.debug()) {
-            builder.client(getDefaultOkHttpClient());
+        // 此处没有保证线程安全，但是无所谓
+        if (mService == null) {
+            Retrofit.Builder builder = new Retrofit.Builder()
+                    .baseUrl(getURL());
+            if (client == null && mHttpConfig.debug()) {
+                builder.client(getDefaultOkHttpClient());
+            }
+            builder.addCallAdapterFactory(RxJava2CallAdapterFactory.createAsync());
+            mService = builder.build().create(HttpService.class);
         }
-        return builder.build().create(HttpService.class);
+        return mService;
     }
 
     private HttpService getDownloadService(OkHttpClient client, String baseUrl, OnProgressUpdateListener listener) {
-        Retrofit.Builder builder = new Retrofit.Builder()
-                .baseUrl(baseUrl == null ? getURL() : baseUrl);
-        OkHttpClient.Builder clientBuilder = null;
-        if (client == null) {
-            clientBuilder = new OkHttpClient.Builder();
-        } else {
-            clientBuilder = client.newBuilder();
+        // 此处没有保证线程安全，但是无所谓
+        if (mDownloadService == null) {
+            Retrofit.Builder builder = new Retrofit.Builder()
+                    .baseUrl(baseUrl == null ? getURL() : baseUrl);
+            OkHttpClient.Builder clientBuilder = null;
+            if (client == null) {
+                clientBuilder = new OkHttpClient.Builder();
+            } else {
+                clientBuilder = client.newBuilder();
+            }
+            builder.client(getDownloadOkHttpClient(clientBuilder, listener));
+            builder.addCallAdapterFactory(RxJava2CallAdapterFactory.createAsync());
+            mDownloadService = builder.build().create(HttpService.class);
         }
-        builder.client(getDownloadOkHttpClient(clientBuilder, listener));
-        return builder.build().create(HttpService.class);
+        return mDownloadService;
     }
 
     private OkHttpClient getDefaultOkHttpClient() {
